change timezone and add a (slow) name to the log view
